take cast into consideration

+++ common/functions/src/scalars/maths/abs_test.rs
@@ -35,7 +35,7 @@ fn test_abs_function() -> Result<()> {
                 Series::new([-1]).into(),
                 DataField::new("arg1", DataType::Int32, false),
             ),
-            expect: Ok(Series::new(vec![1]).into()),
+            expect: Ok(Series::new(vec![1 as u32]).into()),
         },
         Test {
             name: "abs(-10086)",
@@ -44,7 +44,7 @@ fn test_abs_function() -> Result<()> {
                 Series::new([-10086]).into(),
                 DataField::new("arg1", DataType::Int32, false),
             ),
-            expect: Ok(Series::new(vec![10086]).into()),
+            expect: Ok(Series::new(vec![10086 as u32]).into()),
         },
         Test {
             name: "abs('str')",
